requirements.txt
django==6.0.3
djangorestframework==3.17.1
django-filter==25.2
python-dotenv==1.2.2
djangorestframework-simplejwt==5.5.1
requests==2.33.0
psycopg==3.3.3
django-cors-headers==4.9.0
Pillow==12.1.1
django-mptt==0.18.0